Show and Create post page
Create the show and create page. Create a PostLinks component where I put a link to the create post page. Create the Message component. Add a Add and Fail SVG icon component to be in the message component. Create TextArea component. Update the auth hook to return csfr token and the mutate function.